# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "api-server",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/{slug}/threads/{thread_id}": {
      "get": {
        "tags": [
          "Threads"
        ],
        "operationId": "routes_threads_get_thread",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input parameters"
          },
          "404": {
            "description": "Not Found - The requested resource was not found"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Thread": {
        "type": "object",
        "required": [
          "id",
          "last_date",
          "mailing_list_id",
          "root_message_id",
          "start_date",
          "subject"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mailing_list_id": {
            "type": "integer",
            "format": "int32"
          },
          "root_message_id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date-time"
          },
          "last_date": {
            "type": "string",
            "format": "date-time"
          },
          "message_count": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
      "ThreadDetail": {
        "type": "object",
        "required": [
          "emails",
          "thread"
        ],
        "properties": {
          "thread": {
            "$ref": "#/components/schemas/Thread"
          },
          "emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmailHierarchy"
            }
          }
        }
      },
      "EmailHierarchy": {
        "type": "object",
        "required": [
          "author_email",
          "author_id",
          "date",
          "depth",
          "git_commit_hash",
          "id",
          "mailing_list_id",
          "message_id",
          "subject"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mailing_list_id": {
            "type": "integer",
            "format": "int32"
          },
          "message_id": {
            "type": "string"
          },
          "git_commit_hash": {
            "type": "string"
          },
          "author_id": {
            "type": "integer",
            "format": "int32"
          },
          "subject": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "in_reply_to": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "author_name": {
            "type": "string",
            "nullable": true
          },
          "author_email": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}